refactor!: disallow multiple stream filter
